crud user & pubDetails

+++ api/src/controllers/publication/deletePublication.js
@@ -3,10 +3,12 @@ const { Publication } = require('../../db')
 const deletePublication = async(req, res) => {
   const { id } = req.params
   try {
-    await Publication.destroy({
-      where: { id: id }
-    })
-    res.status(200).send(`Publication ${id} has been deleted`)
+    if(id) {
+      await Publication.destroy({
+        where: { id: id }
+      })
+      res.status(200).send(`Publication ${id} has been deleted`)
+    }
   } catch (error) {
     res.status(500).send(error.message)
   }
